requirements.txt
PyQt5>=5.15.0
pyautogui>=0.9.54
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
keyboard>=0.13.5
pyinstaller>=6.0.0
